Fixes for AccountController. Added SecurityDefinitions for swagger.

+++ src/components/responses/account.go
@@ -2,13 +2,13 @@ package responses
 
 type AccountSignUpResponse struct {
 	response
-	Id       int
+	Id       int64
 	Username string
 }
 
 type AccountMeResponse struct {
 	response
-	Id       int
+	Id       int64
 	Username string
 	Balance  float64
 }
